Fix: Remove unnecessary docker compose volume mount
The volume mount for the server in the
Docker Compose file was a holdover from an
experiment to do hot reloading in Docker,
which I later decided against.
It prevents the solution from building cleanly
with docker compose up after a fresh clone,
as it overwrites the solution correctly built
in Docker with an empty folder at start up time.

It was not detected during testing because I
was building from my working copy which
always had a working node_modules directory
from local runs and builds.